+++ kontour/GPage.h
@@ -126,7 +126,7 @@ public:
   void drawContentsInRegion(QPainter &p, const KoRect &r, bool withBasePoints = false, bool outline = false, bool withEditMarks = true);
 
   void invalidateClipRegions();
-  GObject *findContainingObject(int x, int y);
+  GObject *findContainingObject(double x, double y);
   bool findNearestObject(const QString &otype, int x, int y, double max_dist, GObject *&obj, int &pidx, bool all = false);
   bool findContainingObjects(int x, int y, QPtrList<GObject> &olist);
   bool findObjectsContainedIn(const KoRect &r, QPtrList<GObject> &olist);
